Getting Shot Increases Knockback Taken
KnockbackModifier keeps track of your current damage modifier, which
increases when you get shot. That is factored into how much knockback
you take when you get shot. I also added a crate (with a rigidbody but
NOT any of the knockback info players have) to shoot when you're not
shooting the dummy.