Various Fixes
Moved arrow firing to Update() instead of FixedUpdate(), so the
"sticking" bug with releasing arrows should be fixed. Also fixed a bug
where player health would not be reset on death if they're a client, so
you could build up "player 2"s knockback level forever so that a single
arrow sends them flying. Also fixed a bug where the display for health
would not update if they're a client. Also fixed a bug where being
kicked back to the main menu would not give the mouse back.

Next on the agenda: Pause menu and kill counter.